Map to rCRS.
Instead of mapping to nuclear genome to remove all possible NUMTs (this results in very low and inconsistent coverage), MitoSeek will now remap all mitochondrial reads to the rCRS reference genome. 
If the original sequence was aligned to the Yoruba (hg19 and below) reference sequence, it should be remapped to the rCRS for consistency of reporting variants in the literature. 
If the original sequence was aligned to either the hg19a or GRCh38 reference sequences (or anything newer), mitochondrial reads will already be aligned to the rCRS and there will be no need to re-align.